Splitter ut "Allerede lastet opp"
